+++ app/src/main/java/libelulati/tripctrl/Usuarios/UsuarioEditActivity.java
@@ -1,4 +1,4 @@
-package libelulati.tripctrl;
+package libelulati.tripctrl.Usuarios;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -6,6 +6,8 @@ import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
+
+import libelulati.tripctrl.R;
 
 public class UsuarioEditActivity extends AppCompatActivity {
