+++ app/src/main/res/layout/activity_notificacoes_configuracao.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent" android:fitsSystemWindows="true"
-    tools:context="libelulati.tripctrl.NotificacoesConfiguracaoActivity">
+    tools:context="libelulati.tripctrl.Notificacoes.NotificacoesConfiguracaoActivity">
 
     <android.support.design.widget.AppBarLayout android:layout_height="wrap_content"
         android:layout_width="match_parent" android:theme="@style/AppTheme.AppBarOverlay">
