+++ app/src/main/res/layout/content_categoria_list.xml
@@ -8,6 +8,6 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:showIn="@layout/activity_categoria_list"
-    tools:context="libelulati.tripctrl.CategoriaListActivity">
+    tools:context="libelulati.tripctrl.Categorias.CategoriaListActivity">
 
 </RelativeLayout>
