+++ app/src/main/res/layout/content_sobre.xml
@@ -7,6 +7,6 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:showIn="@layout/activity_sobre" tools:context="libelulati.tripctrl.SobreActivity">
+    tools:showIn="@layout/activity_sobre" tools:context="libelulati.tripctrl.Aplicativo.SobreActivity">
 
 </RelativeLayout>
