+++ app/src/main/res/layout/content_tipos_pagamento_list.xml
@@ -8,6 +8,6 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:showIn="@layout/activity_tipos_pagamento_list"
-    tools:context="libelulati.tripctrl.TiposPagamentoListActivity">
+    tools:context="libelulati.tripctrl.TipoPagamento.TiposPagamentoListActivity">
 
 </RelativeLayout>
