+++ app/src/main/res/layout/content_usuario_edit.xml
@@ -8,6 +8,6 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:showIn="@layout/activity_usuario_edit"
-    tools:context="libelulati.tripctrl.UsuarioEditActivity">
+    tools:context="libelulati.tripctrl.Usuarios.UsuarioEditActivity">
 
 </RelativeLayout>
